update RCBase count to 32 bit to avoid overflowing reference counting for ITk

+++ DetectorDescription/GeoModel/GeoModelKernel/GeoModelKernel/RCBase.h
@@ -102,7 +102,7 @@ class RCBase
       //## Attribute: m_count%3CD12A9F013D
       //	The reference count
       //## begin RCBase::m_count%3CD12A9F013D.attr preserve=no  private: mutable unsigned short int {V} 0
-      mutable unsigned short int m_count;
+      mutable unsigned long int m_count;
       //## end RCBase::m_count%3CD12A9F013D.attr
 
     // Additional Private Declarations
